luadns: use ttl defined in the config for the zone
This removes the hardcoded ttl value for zones in the config, and uses
them when updating the record(s) associated with the zone. I picked a
relatively save default of 3600 seconds, which most (?) DNS providers
seem to use for defaults (based on very limited experience, but 1hr
isn't that bad).